+++ .eslintrc.js
@@ -1,0 +1,15 @@
+module.exports = {
+  'env': {
+    'browser': true,
+    'commonjs': true,
+    'es2021': true,
+  },
+  'extends': [
+    'google',
+  ],
+  'parserOptions': {
+    'ecmaVersion': 12,
+  },
+  'rules': {
+  },
+};