+++ api.js
@@ -1,0 +1,58 @@
+/* eslint-disable max-len */
+/*
+
+Copyright (c) 2019 - present AppSeed.us
+
+*/
+const express = require('express');
+const bodyParser = require('body-parser');
+const passport = require('passport');
+const mongoose = require('mongoose');
+const compression = require('compression');
+const https = require('https');
+const http = require('http');
+const fs = require('fs');
+const cors = require('cors');
+const path = require('path');
+const db = require('./config/keys').mongoURI;
+
+require('dotenv').config();
+
+// Instantiate express
+const app = express();
+app.use(compression());
+
+// Passport Config
+require('./config/passport')(passport);
+
+// DB Config
+
+// Connect to MongoDB
+mongoose
+    .connect(
+        db, {useNewUrlParser: true,
+          useFindAndModify: false,
+          useUnifiedTopology: true,
+          useCreateIndex: true},
+    )
+    .then(() => console.log('MongoDB Connected'))
+    .catch((err) => console.log(err));
+
+app.use(cors());
+
+// Express body parser
+app.use('/public', express.static('public'));
+app.use(bodyParser.urlencoded({extended: true}));
+app.use(bodyParser.json());
+
+// Initialize routes middleware
+app.use('/api/users', require('./routes/users'));
+
+const PORT = process.env.PORT;
+
+http.createServer({
+}, app)
+    .listen(PORT, function() {
+      console.log('App listening on port ' + PORT + '! Go to http://localhost:' + PORT + '/');
+    });
+