+++ config/config.js
@@ -1,0 +1,9 @@
+/* eslint-disable max-len */
+/*
+
+Copyright (c) 2019 - present AppSeed.us
+
+*/
+module.exports = {
+  webURL: 'https://localhost:5000/',
+};