+++ config/keys.js
@@ -1,0 +1,12 @@
+/* eslint-disable max-len */
+/*
+
+Copyright (c) 2019 - present AppSeed.us
+
+*/
+dbPasswordDev = 'mongodb://localhost/api_server_nodejs';
+
+module.exports = {
+  mongoURI: dbPasswordDev,
+  secret: 'SuperS3cret_4277m',
+};